LiteraSocial-main/LiteraSocial-main/Backend/requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
git-filter-repo==2.47.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
motor==3.7.1
mpmath==1.3.0
networkx==3.4.2
numpy
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-stubs==0.0.11
opt_einsum==3.4.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
premailer==3.10.0
protobuf==4.25.6
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.6
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yagmail==0.15.293
